feat(rk3588): introduce SD/MMC Control Service
The whole point of this service is to allow the Windows SD/eMMC drivers
to effortlessly control necessary clocks and voltages, since this is all
done in non-standard ways via CRU and a PMIC on the SPI bus.

Why not use ACPI methods?

We want to be able to collect crash dumps when SD/eMMC is used as a boot
drive. The crash dump stack is much more limited and does not provide
any means to call into the ACPI driver.

Change-Id: I2b51f787a5255b2a30a6a5ec14f00031f351e6b2
Signed-off-by: Mario Bălănică <mariobalanica02@gmail.com>